requirements.txt
#pip install -r requirements.txt --default-timeout=120
requests>=2.25.1
ffmpeg-python
imageio-ffmpeg
moviepy
opencv-python==4.9.0.80
python-dotenv
aiofiles==23.2.1
aiohttp==3.9.1
anyio==4.2.0
attrs==23.1.0
azure-core>=1.31.0
azure-eventhub==5.11.5
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-servicebus==7.11.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Flask==3.0.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-speech==2.23.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
moviepy==1.0.3
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.4
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sounddevice==0.4.6
SpeechRecognition==3.10.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
youtube-transcript-api==0.6.1
python-dotenv
